+++ projects/wotw-server/src/jvmMain/kotlin/wotw/server/util/Helpers.kt
@@ -23,6 +23,6 @@ inline fun <reified R : Any> Route.put(
 }
 
 fun rezero(n: Int) = if(n == -1) 0 else n
-fun rezero(n: Float) = if(n == -1f) 0f else n
+fun rezero(n: Double) = if(n == -1.0) 0.0 else n
 fun zerore(n: Int) = if(n == 0) -1 else n
-fun zerore(n: Float) = if(n == 0f) -1f else n
+fun zerore(n: Double) = if(n == 0.0) -1.0 else n
